services/bq_hot_loader/requirements.txt
google-cloud-pubsub
google-cloud-bigquery
jsonschema
